Fix customer screen

+++ Controls/usrJob.cs
@@ -1468,6 +1468,9 @@ namespace Astrodon.Controls
             string outputFileName = "JobProcessing_" + DateTime.Now.ToString("yyyyMMdd HHmmss") + ".pdf";
             string desktopFolder = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
 
+            if (statementFileList.Count() <= 0)
+                return;
+
             outputFileName = Path.Combine(desktopFolder, outputFileName);
             if (File.Exists(outputFileName))
                 File.Delete(outputFileName);
